# CMake simplification & build speedup