--- a/cmake/stdcxx_reduce_exports.ver
+++ b/cmake/stdcxx_reduce_exports.ver
@@ -1,7 +1,0 @@
-{
-local:
-extern "C++"
-{
-        std::locale::id*;
-};
-};