--- a/src/CMakeLists.txt
+++ b/src/CMakeLists.txt
@@ -18,7 +18,6 @@ subdirs(
   globals
   image
   lambda
-  lazy_eval
   logger
   ml_data
   nanosockets