--- a/src/capi/impl/capi_visualization.cpp
+++ b/src/capi/impl/capi_visualization.cpp
@@ -113,6 +113,7 @@ EXPORT bool tc_plot_finished_streaming(const tc_plot* plot,
 }
 
 #ifdef __APPLE__
+#ifndef TC_BUILD_IOS
 EXPORT void tc_plot_render_final_into_context(const tc_plot* plot,
                                               tc_plot_variation variation,
                                               CGContextRef context,
@@ -144,6 +145,7 @@ EXPORT bool tc_plot_render_next_into_context(const tc_plot* plot,
   ERROR_HANDLE_END(error, true);
 
 }
-#endif
+#endif // TC_BUILD_IOS
+#endif // __APPLE__
 
 } // extern "C"