--- a/src/unity/toolkits/coreml_export/MLModel/build/format/CategoricalMapping_enums.h
+++ b/src/unity/toolkits/coreml_export/MLModel/build/format/CategoricalMapping_enums.h
@@ -8,6 +8,7 @@ enum MLCategoricalMappingMappingType: int {
     MLCategoricalMappingMappingType_NOT_SET = 0,
 };
 
+__attribute__((__unused__))
 static const char * MLCategoricalMappingMappingType_Name(MLCategoricalMappingMappingType x) {
     switch (x) {
         case MLCategoricalMappingMappingType_stringToInt64Map:
@@ -25,6 +26,7 @@ enum MLCategoricalMappingValueOnUnknown: int {
     MLCategoricalMappingValueOnUnknown_NOT_SET = 0,
 };
 
+__attribute__((__unused__))
 static const char * MLCategoricalMappingValueOnUnknown_Name(MLCategoricalMappingValueOnUnknown x) {
     switch (x) {
         case MLCategoricalMappingValueOnUnknown_strValue: