--- a/src/unity/toolkits/coreml_export/MLModel/build/format/CustomModel_enums.h
+++ b/src/unity/toolkits/coreml_export/MLModel/build/format/CustomModel_enums.h
@@ -12,6 +12,7 @@ enum MLCustomModelParamValuevalue: int {
     MLCustomModelParamValuevalue_NOT_SET = 0,
 };
 
+__attribute__((__unused__))
 static const char * MLCustomModelParamValuevalue_Name(MLCustomModelParamValuevalue x) {
     switch (x) {
         case MLCustomModelParamValuevalue_doubleValue: