--- a/src/unity/toolkits/coreml_export/MLModel/build/format/DictVectorizer_enums.h
+++ b/src/unity/toolkits/coreml_export/MLModel/build/format/DictVectorizer_enums.h
@@ -8,6 +8,7 @@ enum MLDictVectorizerMap: int {
     MLDictVectorizerMap_NOT_SET = 0,
 };
 
+__attribute__((__unused__))
 static const char * MLDictVectorizerMap_Name(MLDictVectorizerMap x) {
     switch (x) {
         case MLDictVectorizerMap_stringToIndex: