--- a/src/unity/toolkits/coreml_export/MLModel/build/format/FeatureTypes_enums.h
+++ b/src/unity/toolkits/coreml_export/MLModel/build/format/FeatureTypes_enums.h
@@ -15,6 +15,7 @@ enum MLImageFeatureTypeSizeFlexibility: int {
     MLImageFeatureTypeSizeFlexibility_NOT_SET = 0,
 };
 
+__attribute__((__unused__))
 static const char * MLImageFeatureTypeSizeFlexibility_Name(MLImageFeatureTypeSizeFlexibility x) {
     switch (x) {
         case MLImageFeatureTypeSizeFlexibility_enumeratedSizes:
@@ -39,6 +40,7 @@ enum MLArrayFeatureTypeShapeFlexibility: int {
     MLArrayFeatureTypeShapeFlexibility_NOT_SET = 0,
 };
 
+__attribute__((__unused__))
 static const char * MLArrayFeatureTypeShapeFlexibility_Name(MLArrayFeatureTypeShapeFlexibility x) {
     switch (x) {
         case MLArrayFeatureTypeShapeFlexibility_enumeratedShapes:
@@ -56,6 +58,7 @@ enum MLDictionaryFeatureTypeKeyType: int {
     MLDictionaryFeatureTypeKeyType_NOT_SET = 0,
 };
 
+__attribute__((__unused__))
 static const char * MLDictionaryFeatureTypeKeyType_Name(MLDictionaryFeatureTypeKeyType x) {
     switch (x) {
         case MLDictionaryFeatureTypeKeyType_int64KeyType:
@@ -73,6 +76,7 @@ enum MLSequenceFeatureTypeType: int {
     MLSequenceFeatureTypeType_NOT_SET = 0,
 };
 
+__attribute__((__unused__))
 static const char * MLSequenceFeatureTypeType_Name(MLSequenceFeatureTypeType x) {
     switch (x) {
         case MLSequenceFeatureTypeType_int64Type:
@@ -95,6 +99,7 @@ enum MLFeatureTypeType: int {
     MLFeatureTypeType_NOT_SET = 0,
 };
 
+__attribute__((__unused__))
 static const char * MLFeatureTypeType_Name(MLFeatureTypeType x) {
     switch (x) {
         case MLFeatureTypeType_int64Type: