--- a/src/unity/toolkits/coreml_export/MLModel/build/format/GLMClassifier_enums.h
+++ b/src/unity/toolkits/coreml_export/MLModel/build/format/GLMClassifier_enums.h
@@ -18,6 +18,7 @@ enum MLGLMClassifierClassLabels: int {
     MLGLMClassifierClassLabels_NOT_SET = 0,
 };
 
+__attribute__((__unused__))
 static const char * MLGLMClassifierClassLabels_Name(MLGLMClassifierClassLabels x) {
     switch (x) {
         case MLGLMClassifierClassLabels_stringClassLabels: