--- a/src/unity/toolkits/coreml_export/MLModel/build/format/Model_enums.h
+++ b/src/unity/toolkits/coreml_export/MLModel/build/format/Model_enums.h
@@ -33,6 +33,7 @@ enum MLModelType: int {
     MLModelType_NOT_SET = 0,
 };
 
+__attribute__((__unused__))
 static const char * MLModelType_Name(MLModelType x) {
     switch (x) {
         case MLModelType_pipelineClassifier: