--- a/src/unity/toolkits/coreml_export/MLModel/build/format/SVM_enums.h
+++ b/src/unity/toolkits/coreml_export/MLModel/build/format/SVM_enums.h
@@ -10,6 +10,7 @@ enum MLKernelkernel: int {
     MLKernelkernel_NOT_SET = 0,
 };
 
+__attribute__((__unused__))
 static const char * MLKernelkernel_Name(MLKernelkernel x) {
     switch (x) {
         case MLKernelkernel_linearKernel:
@@ -31,6 +32,7 @@ enum MLSupportVectorRegressorsupportVectors: int {
     MLSupportVectorRegressorsupportVectors_NOT_SET = 0,
 };
 
+__attribute__((__unused__))
 static const char * MLSupportVectorRegressorsupportVectors_Name(MLSupportVectorRegressorsupportVectors x) {
     switch (x) {
         case MLSupportVectorRegressorsupportVectors_sparseSupportVectors:
@@ -48,6 +50,7 @@ enum MLSupportVectorClassifiersupportVectors: int {
     MLSupportVectorClassifiersupportVectors_NOT_SET = 0,
 };
 
+__attribute__((__unused__))
 static const char * MLSupportVectorClassifiersupportVectors_Name(MLSupportVectorClassifiersupportVectors x) {
     switch (x) {
         case MLSupportVectorClassifiersupportVectors_sparseSupportVectors:
@@ -65,6 +68,7 @@ enum MLSupportVectorClassifierClassLabels: int {
     MLSupportVectorClassifierClassLabels_NOT_SET = 0,
 };
 
+__attribute__((__unused__))
 static const char * MLSupportVectorClassifierClassLabels_Name(MLSupportVectorClassifierClassLabels x) {
     switch (x) {
         case MLSupportVectorClassifierClassLabels_stringClassLabels: