--- a/src/unity/toolkits/coreml_export/MLModel/build/format/TextClassifier_enums.h
+++ b/src/unity/toolkits/coreml_export/MLModel/build/format/TextClassifier_enums.h
@@ -7,6 +7,7 @@ enum MLTextClassifierClassLabels: int {
     MLTextClassifierClassLabels_NOT_SET = 0,
 };
 
+__attribute__((__unused__))
 static const char * MLTextClassifierClassLabels_Name(MLTextClassifierClassLabels x) {
     switch (x) {
         case MLTextClassifierClassLabels_stringClassLabels: