--- a/src/unity/toolkits/coreml_export/mlmodel_include.hpp
+++ b/src/unity/toolkits/coreml_export/mlmodel_include.hpp
@@ -9,6 +9,7 @@
 #if defined(__GNUC__)
   #pragma GCC diagnostic push
   #pragma GCC diagnostic ignored "-Wreturn-type"
+  #pragma GCC diagnostic ignored "-Wunused-function"
 #elif defined (__clang__)
   #pragma clang diagnostic push
   #pragma clang diagnostic ignored "-Wsign-conversion"