--- a/test/optimization/optimization_interface_test.cxx
+++ b/test/optimization/optimization_interface_test.cxx
@@ -21,7 +21,7 @@
 #include <optimization/newton_method-inl.hpp>
 #include <optimization/gradient_descent-inl.hpp>
 #include <optimization/accelerated_gradient-inl.hpp>
-#include <optimization/lbfgs.hpp>
+#include <optimization/lbfgs-inl.hpp>
 
 
 using namespace turi;