… src/capi/CMakeLists.txt
… src/capi/framework/tc_capi_exports_osx.ver
… src/cppipc/CMakeLists.txt
… src/dot_graph_printer/CMakeLists.txt
… src/external/CMakeLists.txt
… src/external/aws-sdk-cpp/CMakeLists.txt
… src/external/aws-sdk-cpp/aws-cpp-sdk-core/CMakeLists.txt
… src/external/aws-sdk-cpp/aws-cpp-sdk-s3/CMakeLists.txt
… src/external/google/protobuf/CMakeLists.txt
… src/external/libjson/CMakeLists.txt
… src/external/lz4/CMakeLists.txt
… src/external/sparsehash/CMakeLists.txt
… src/external/uuid_p/CMakeLists.txt
… src/external/xgboost/CMakeLists.txt
… src/fileio/CMakeLists.txt
… src/flexible_type/CMakeLists.txt
… src/globals/CMakeLists.txt
… src/image/CMakeLists.txt
… src/lambda/CMakeLists.txt
… src/lazy_eval/CMakeLists.txt
… src/logger/CMakeLists.txt
… src/ml_data/CMakeLists.txt
… src/nanosockets/CMakeLists.txt
… src/objc/CMakeLists.txt
… src/optimization/CMakeLists.txt
… src/platform/CMakeLists.txt
… src/platform/config/CMakeLists.txt
… src/platform/crash_handler/CMakeLists.txt
… src/platform/minipsutil/CMakeLists.txt
… src/platform/network/CMakeLists.txt