The general idea behind these changes is to prevent linking targets that
aren't actually shipped from this repo. Prior to this PR, the repo
contains quite a few static libraries that aren't shipped anywhere (and
may or may not be included in various other targets), which result in an
extra cost (to CPU and disk) both in assembling and preserving the `.a` files.

This change reorients building and linking along a different axis.

Previously: each module within the build (typically a single folder
under `src/`) would get compiled and linked into its own static library.
Those static libraries would then get (usually) get linked into
libunity_shared.so or another Python lib, another library for execution
(libTuriCreate/libRecommender/etc.), or a test executable.

With this PR: each module within the build only gets compiled (into a
`OBJECT` library, which is a way of tracking collections of `.o` files
as a single library in CMake). Each final link target (i.e.
libunity_shared.so) then compiles with all the .o files directly
(essentially bypassing compilation and just linking instead). This also
allows us to get rid of the usage of things like `make_shared_from_static`.

Bonus change: all tests now link against
`libunity_shared_for_testing.so` instead of statically linking all their
dependencies, so they are much smaller now.

Disk usage (debug, macOS):

* Before: 17.0 GB
* After: 4.3 GB

Build time (debug, 15" 2015 MBP, primed ccache):

* Before: 927.47s user 487.59s system 318% cpu 7:23.84 total
* After: 575.39s user 406.67s system 256% cpu 6:22.35 total

Build time (Travis CI, approximate):

* Before: About 2 hours total (including running tests).
* After: About 1:51 total (including running tests).