@ylow 2019-02-15
A suggestion is to keep the original make_library macro but making it just a thin wrapper around "add_library( ... OBJECT ...)". We can then maintain the original REQUIRE flags in all the CMakeLists.txt, but we just ignore the REQUIRE arguments until we upgrade to a newer CMake which can handle the transitive dependencies correctly.